requirements.txt
anyio==3.5.0
asgiref==3.5.0
black==22.3.0
click==8.0.4
fastapi==0.75.0
flake8==4.0.1
h11==0.13.0
httptools==0.3.0
idna==3.3
Jinja2==3.0.3
MarkupSafe==2.1.0
psycopg2-binary==2.9.3
pydantic==1.9.0
python-dotenv==0.19.2
PyYAML==6.0
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.1.1
uvicorn==0.18.2
watchgod==0.7
websockets==10.2